requirements.txt
leverj-ordersigner==0.9.1
requests==2.24.0
websocket_client==0.57.0
aiohttp
python-socketio[asyncio_client]